[refactor/slices] Flow slice
