+++ src/store/game/flow/flow.slice.ts
@@ -2,13 +2,13 @@ import { EStatus } from "@/shared/enums/status.enum";
 import { IFlowSlice, TFlowSliceCreator } from "./flow.types";
 
 export const createFlowSlice: TFlowSliceCreator<IFlowSlice> = (set, get) => ({
-  status: EStatus.SELECT_ACTION,
+  gameStatus: EStatus.SELECT_ACTION,
   round: 1,
 
-  setStatus: (status) => {
-    set({ status });
-  },
+  // --- SETTERS ---
+  setGameStatus: (gameStatus) => set({ gameStatus }),
 
+  // --- METHODS ---
   nextRound: () => {
     set((state) => ({ round: state.round + 1 }));
   },
