+++ src/store/game/flow/flow.types.ts
@@ -2,9 +2,13 @@ import { StateCreator } from "zustand";
 import { EStatus } from "@/shared/enums/status.enum";
 
 export interface IFlowSlice {
-  status: EStatus;
+  gameStatus: EStatus;
   round: number;
-  setStatus: (status: EStatus) => void;
+
+  // --- SETTERS ---
+  setGameStatus: (gameStatus: EStatus) => void;
+
+  // --- METHODS ---
   nextRound: () => void;
 }
