+++ src/store/game/flow/tests/setup/create-test-store.ts
@@ -1,0 +1,7 @@
+import { create } from "zustand";
+import { createFlowSlice, IFlowSlice } from "../..";
+
+export const createTestStore = () =>
+  create<IFlowSlice>((...args) => ({
+    ...createFlowSlice(...args),
+  }));
